Move all pytest code to nltk/test/ (#2697)
* Move all pytest code to nltk/test

* pull out all test code from non-test source
