+++ nltk/stem/wordnet.py
@@ -40,10 +40,3 @@ class WordNetLemmatizer(object):
 
     def __repr__(self):
         return "<WordNetLemmatizer>"
-
-
-# unload wordnet
-def teardown_module(module=None):
-    from nltk.corpus import wordnet
-
-    wordnet._unload()
