+++ nltk/test/metrics.doctest
@@ -5,6 +5,14 @@
 Metrics
 =======
 
+-----
+Setup
+-----
+
+    >>> import pytest
+    >>> _ = pytest.importorskip("numpy")
+
+
 The `nltk.metrics` package provides a variety of *evaluation measures*
 which can be used for a wide variety of NLP tasks.
